registeration done image upload in cloudinary and production grade postman configuration done

+++ src/utils/cloudinary.js
@@ -1,28 +1,33 @@
-import {v2 as cloudinary} from "cloudinary";
+import { v2 as cloudinary } from "cloudinary";
 import fs from "fs";
 
 
-    cloudinary.config({ 
-        cloud_name: process.env.CLOUDINARY_CLOUD_NAME, 
-        api_key: process.env.CLOUDINARY_API_KEY, 
-        api_secret: process.env.CLOUDINARY_API_SECRET
-    });
+cloudinary.config({
+    cloud_name: process.env.CLOUDINARY_CLOUD_NAME,
+    api_key: process.env.CLOUDINARY_API_KEY,
+    api_secret: process.env.CLOUDINARY_API_SECRET
+});
 
-    const uploadOnCloudinary = async(localFilePath) =>{
-           try{
-             if(!localFilePath) return null
-             // const uploadResult = await cloudinary.uploader
-            const res = await  cloudinary.uploader.upload
-            (localFilePath , {
-                resource_type : "auto"
+const uploadOnCloudinary = async (localFilePath) => {
+    try {
+        if (!localFilePath) return null
+        // const uploadResult = await cloudinary.uploader
+        const res = await cloudinary.uploader.upload
+            (localFilePath, {
+                resource_type: "auto"
             })
-            // file has been uploaded successfull
-            console.log("file is upload on cloudinary" ,  response.url);
-           
-           }catch(err){
-             fs.unlinkSync(localFilePath)
-              return null
-           }
-    }
+        // file has been uploaded successfull
+        // console.log("file is upload on cloudinary", res.url);
+        fs.unlinkSync(localFilePath);
+        return res;
 
-    export {uploadOnCloudinary}
+    } catch (err) {
+        console.error("Cloudinary Error:", err);
+        if (fs.existsSync(localFilePath)) {
+            fs.unlinkSync(localFilePath);
+        }
+        return null
+    }
+}
+
+export { uploadOnCloudinary }
